# Speed up list by using compacted strings