--- a/src/Common/CompactStrings.cpp
+++ b/src/Common/CompactStrings.cpp
@@ -1,6 +1,4 @@
-#include "CompactStrings.h"
-
-#include "Service/memcopy.h"
+#include <Common/CompactStrings.h>
 
 
 namespace RK
@@ -22,17 +20,6 @@ void CompactStrings::reserve(size_t n, size_t total_size_)
         data.reserve(total_size_);
     else
         data.reserve(AVG_ELEMENT_SIZE_HINT * total_size_);
-}
-
-void CompactStrings::push_back(const String & s)
-{
-    const size_t old_size = data.size();
-    const size_t size_to_append = s.size();
-    const size_t new_size = old_size + size_to_append;
-
-    data.resize(new_size);
-    memcopy(data.data() + old_size, s.c_str(), size_to_append);
-    offsets.push_back(new_size);
 }
 
 StringRef CompactStrings::operator[](int64_t i) const