--- a/src/Common/CompactStrings.h
+++ b/src/Common/CompactStrings.h
@@ -1,7 +1,8 @@
 #pragma once
 
-#include "PODArray.h"
-#include "common/StringRef.h"
+#include <Common/PODArray.h>
+#include <Service/memcopy.h>
+#include <common/StringRef.h>
 
 
 namespace RK
@@ -48,7 +49,17 @@ public:
 
     void reserve(size_t n, size_t total_size = 0);
 
-    void push_back(const String & s);
+    inline void push_back(const String & s)
+    {
+        const size_t old_size = data.size();
+        const size_t size_to_append = s.size();
+        const size_t new_size = old_size + size_to_append;
+
+        data.resize(new_size);
+        memcopy(data.data() + old_size, s.c_str(), size_to_append);
+        offsets.push_back(new_size);
+    }
+
     template <class Ttr> void push_back(Ttr begin, Ttr end)
     {
         for (Ttr it = begin; it != end; ++it)