--- a/src/Service/memcopy.h
+++ b/src/Service/memcopy.h
@@ -16,39 +16,53 @@ namespace RK
 {
 #ifdef __SSE2__
 
-inline void memcopy(char * __restrict dst, const char * __restrict src, size_t n)
+/// 1. src and dst has right padding which is not less than 15 bytes, or
+/// 2. n is multiple of 16
+inline void memcopyWithRightPadding15(char * __restrict dst, const char * __restrict src, size_t n)
 {
-    auto aligned_n = n / 16 * 16;
-    auto left = n - aligned_n;
-    while (aligned_n > 0)
+    while (n > 0)
     {
         _mm_storeu_si128(reinterpret_cast<__m128i *>(dst), _mm_loadu_si128(reinterpret_cast<const __m128i *>(src)));
 
         dst += 16;
         src += 16;
-        aligned_n -= 16;
+        n -= 16;
 
         /// Avoid clang loop-idiom optimization, which transforms _mm_storeu_si128 to built-in memcpy
         __asm__ __volatile__("" : : : "memory");
     }
-    ::memcpy(dst, src, left);
 }
-
-#elif defined(__aarch64__) && defined(__ARM_NEON)
 
 inline void memcopy(char * __restrict dst, const char * __restrict src, size_t n)
 {
     auto aligned_n = n / 16 * 16;
     auto left = n - aligned_n;
-    while (aligned_n > 0)
+    memcopyWithRightPadding15(dst, src, aligned_n);
+    ::memcpy(dst + aligned_n, src + aligned_n, left);
+}
+
+#elif defined(__aarch64__) && defined(__ARM_NEON)
+
+/// 1. src and dst has right padding which is not less than 15 bytes, or
+/// 2. n is multiple of 16
+inline void memcopyWithRightPadding15(char * __restrict dst, const char * __restrict src, size_t n)
+{
+    while (n > 0)
     {
         vst1q_s8(reinterpret_cast<signed char *>(dst), vld1q_s8(reinterpret_cast<const signed char *>(src)));
 
         dst += 16;
         src += 16;
-        aligned_n -= 16;
+        n -= 16;
     }
-    ::memcpy(dst, src, left);
+}
+
+inline void memcopy(char * __restrict dst, const char * __restrict src, size_t n)
+{
+    auto aligned_n = n / 16 * 16;
+    auto left = n - aligned_n;
+    memcopyWithRightPadding15(dst, src, aligned_n);
+    ::memcpy(dst + aligned_n, src + aligned_n, left);
 }
 
 #else