@JackyWoo 2024-06-12
The following is the flame graph when benchmarking. We can see after optimization list request processing takes less CPU cycles.

1. before optimization

![image](https://github.com/JDRaftKeeper/RaftKeeper/assets/3991709/3212fd9c-5795-43b2-9011-1e7050cc3b03)

2. after optimization
![image](https://github.com/JDRaftKeeper/RaftKeeper/assets/3991709/fb3fff98-c4c3-44b0-b152-9e08dc9954f4)